1.超级管理员，权限判断直接通过。2.将模块排序码由int改为double。3.Object类型添加DeepClone扩展方法

+++ src/OSharp.Permissions/Security/Dtos/ModuleBaseInputDto.cs
@@ -34,7 +34,7 @@ namespace OSharp.Core.Security.Dtos
         /// <summary>
         /// 获取或设置 排序码
         /// </summary>
-        public int OrderCode { get; set; }
+        public double OrderCode { get; set; }
 
         /// <summary>
         /// 获取或设置 父模块编号
